+++ src/Smartstore.Core/Migrations/SmartDbContextDataSeeder.cs
@@ -27,6 +27,20 @@ namespace Smartstore.Core.Data.Migrations
             builder.AddOrUpdate("Admin.Customers.RemoveAffiliateAssignment",
                 "Remove assignment to affiliate",
                 "Zuordnung zum Partner entfernen");
+
+            builder.AddOrUpdate("Admin.Configuration.Settings.Order.MaxMessageOrderAgeInDays",
+                "Maximum order age for sending messages",
+                "Maximale Auftragsalter für den Nachrichtenversand",
+                "Specifies the maximum order age in days up to which to create and send messages. Set to 0 to always send messages.",
+                "Legt das maximale Auftragsalter in Tagen fest, bis zu dem Nachrichten erstellt und gesendet werden sollen. Setzen Sie diesen Wert auf 0, um Nachrichten immer zu versenden.");
+
+            builder.AddOrUpdate("Admin.MessageTemplate.OrderTooOldForMessageInfo",
+                "The message \"{0}\" for order {1} was not created. It is too old ({2}).",
+                "Die Nachricht \"{0}\" für Auftrag {1} wurde nicht erstellt. Er ist zu alt ({2}).");
+
+            // Typo.
+            builder.AddOrUpdate("Admin.Configuration.Settings.ShoppingCart.ShowConfirmOrderLegalHint.Hint")
+                .Value("de", "Legt fest, ob rechtliche Hinweise in der Warenkorbübersicht auf der Bestellabschlussseite angezeigt werden. Dieser Text kann in den Sprachresourcen geändert werden.");
         }
     }
 }